feat: xml에서 사용가능한 이미지 피커 생성
