+++ presentation/src/main/kotlin/com/sooum/where_android/view/common/modal/ImagePickerDialog.kt
@@ -1,0 +1,211 @@
+package com.sooum.where_android.view.common.modal
+
+import android.os.Bundle
+import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.ViewCompositionStrategy
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import androidx.fragment.app.DialogFragment
+import com.sooum.domain.model.ImageAddType
+import com.sooum.where_android.R
+import com.sooum.where_android.theme.GrayScale800
+import com.sooum.where_android.theme.pretendard
+
+
+/**
+ * 기본 커버 혹은 이미지 선택
+ */
+@Composable
+fun ImagePickerDialog(
+    onDismiss: () -> Unit,
+    updateImageType: (ImageAddType) -> Unit
+) {
+    Dialog(
+        onDismissRequest = onDismiss
+    ) {
+        Box(
+            Modifier
+                .fillMaxSize()
+                .clickable(
+                    onClick = onDismiss,
+                    indication = null,
+                    interactionSource = remember { MutableInteractionSource() }
+                )
+        ) {
+            ImagePickerDialogContent(
+                Modifier.align(Alignment.Center),
+                onDismiss,
+                updateImageType
+            )
+        }
+    }
+}
+
+/**
+ * Xml 호환용 이미지 피커
+ * ```
+ * ImagePickerDialogFragment.getInstance(this).show(
+ *     parentFragmentManager,ImagePickerDialogFragment.TAG
+ * )
+ * ```
+ */
+class ImagePickerDialogFragment : DialogFragment() {
+
+    interface ImageTypeHandler {
+        fun receiveImageType(imageType: ImageAddType)
+    }
+    private var imageTypeHandler : ImageTypeHandler? = null
+
+    fun setHandler(imageTypeHandler: ImageTypeHandler) {
+        this.imageTypeHandler = imageTypeHandler
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return ComposeView(requireContext()).apply {
+            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
+            setContent {
+                ImagePickerDialogContent(
+                    modifier = Modifier,
+                    onDismiss = {
+                        this@ImagePickerDialogFragment.dismiss()
+                    },
+                    updateImageType = {imageType ->
+                        Log.d("JWH2",imageType.toString())
+                        Log.d("JWH2",imageTypeHandler.toString())
+                        imageTypeHandler?.receiveImageType(imageType)
+                    }
+                )
+            }
+        }
+    }
+
+    companion object {
+        fun getInstance(handler: ImageTypeHandler): ImagePickerDialogFragment {
+            return ImagePickerDialogFragment()
+                .apply {
+                    setHandler(handler)
+                }
+        }
+        const val TAG = "ImagePickerDialogFragment"
+    }
+
+}
+
+@Composable
+private fun ImagePickerDialogContent(
+    modifier: Modifier,
+    onDismiss: () -> Unit,
+    updateImageType: (ImageAddType) -> Unit
+) {
+    val pickSingleMedia =
+        rememberLauncherForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri ->
+            if (uri != null) {
+                updateImageType(ImageAddType.Content(uri))
+                onDismiss()
+            }
+        }
+    Column(
+        modifier = Modifier
+            .width(310.dp)
+            .height(150.dp)
+            .clip(RoundedCornerShape(12.dp))
+            .background(Color.White)
+            .then(modifier),
+        verticalArrangement = Arrangement.Center
+    ) {
+        val textList = listOf(
+            "기본 커버 선택",
+            "앨범에서 사진 선택"
+        )
+        val imageRes = listOf(
+            R.drawable.icon_pick_default,
+            R.drawable.icon_pick_image
+        )
+        val actionList = listOf(
+            {
+                updateImageType(ImageAddType.Default)
+                onDismiss()
+            },
+            {
+                pickSingleMedia.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
+            }
+        )
+
+        repeat(2) {
+            val text = textList[it]
+            val res = imageRes[it]
+            val action = actionList[it]
+            Button(
+                onClick = action,
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = Color.White,
+                    contentColor = GrayScale800
+                )
+            ) {
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.Center,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Image(
+                        painter = painterResource(res),
+                        contentDescription = null
+                    )
+                    Spacer(Modifier.width(4.dp))
+                    Text(
+                        text = text,
+                        fontFamily = pretendard,
+                        fontWeight = FontWeight.Medium,
+                        fontSize = 14.sp
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+@Preview
+private fun ImagePickerDialogContentPreview() {
+    ImagePickerDialogContent(
+        Modifier,{},{}
+    )
+}
