+++ presentation/src/main/res/drawable/icon_pick_image.xml
@@ -1,10 +1,10 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="14dp"
-    android:height="14dp"
-    android:viewportWidth="14"
-    android:viewportHeight="14">
+    android:width="16dp"
+    android:height="16dp"
+    android:viewportWidth="16"
+    android:viewportHeight="16">
   <path
-      android:pathData="M1,16C1,15.936 1,15.869 1,15.8V4.2C1,3.08 1,2.52 1.218,2.092C1.41,1.715 1.715,1.41 2.092,1.218C2.52,1 3.08,1 4.2,1H15.8C16.92,1 17.48,1 17.908,1.218C18.284,1.41 18.59,1.715 18.782,2.092C19,2.519 19,3.079 19,4.197V15.803C19,16.288 19,16.668 18.982,16.977M1,16C1.001,16.988 1.013,17.506 1.218,17.907C1.41,18.284 1.715,18.59 2.092,18.782C2.519,19 3.079,19 4.197,19H15.804C16.921,19 17.48,19 17.908,18.782C18.284,18.59 18.59,18.284 18.782,17.907C18.906,17.665 18.959,17.381 18.982,16.977M1,16L5.768,10.438L5.769,10.436C6.192,9.943 6.404,9.696 6.655,9.606C6.876,9.528 7.117,9.53 7.336,9.611C7.587,9.704 7.795,9.954 8.212,10.454L10.883,13.66C11.269,14.123 11.463,14.355 11.699,14.449C11.906,14.531 12.136,14.541 12.35,14.477C12.594,14.405 12.809,14.19 13.239,13.761L13.736,13.264C14.173,12.826 14.392,12.608 14.64,12.536C14.857,12.473 15.09,12.487 15.299,12.573C15.537,12.672 15.73,12.912 16.117,13.396L18.982,16.977M18.982,16.977L19,17M13,7C12.448,7 12,6.552 12,6C12,5.448 12.448,5 13,5C13.552,5 14,5.448 14,6C14,6.552 13.552,7 13,7Z"
+      android:pathData="M2,12C2,11.957 2,11.913 2,11.867V4.133C2,3.387 2,3.013 2.145,2.728C2.273,2.477 2.477,2.273 2.728,2.145C3.013,2 3.387,2 4.133,2H11.867C12.613,2 12.987,2 13.272,2.145C13.523,2.273 13.727,2.477 13.855,2.728C14,3.013 14,3.386 14,4.131V11.869C14,12.192 14,12.445 13.988,12.652M2,12C2.001,12.659 2.009,13.004 2.145,13.272C2.273,13.523 2.477,13.727 2.728,13.855C3.013,14 3.386,14 4.131,14H11.869C12.614,14 12.987,14 13.272,13.855C13.523,13.727 13.727,13.523 13.855,13.272C13.937,13.11 13.973,12.921 13.988,12.652M2,12L5.179,8.292L5.18,8.291C5.462,7.962 5.603,7.797 5.77,7.738C5.917,7.685 6.078,7.687 6.224,7.741C6.391,7.803 6.53,7.969 6.808,8.303L8.589,10.44C8.846,10.748 8.975,10.904 9.132,10.966C9.271,11.021 9.424,11.027 9.567,10.985C9.729,10.937 9.873,10.794 10.159,10.507L10.491,10.176C10.782,9.884 10.928,9.738 11.093,9.691C11.238,9.649 11.393,9.658 11.533,9.715C11.691,9.781 11.82,9.942 12.078,10.264L13.988,12.652M13.988,12.652L14,12.666M10,6C9.632,6 9.333,5.702 9.333,5.333C9.333,4.965 9.632,4.667 10,4.667C10.368,4.667 10.667,4.965 10.667,5.333C10.667,5.702 10.368,6 10,6Z"
       android:strokeLineJoin="round"
       android:strokeWidth="1.5"
       android:fillColor="#00000000"
